+++ CHANGELOG.md
@@ -9,6 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added a `source` member to `Tileset`, `Map` and `Template`, which stores the resource path they have been loaded from.
 
+## [Unreleased (0.12.2)]
+### Fixed
+- Fixed template instance size and position overrides in `ObjectData::shape`. (#309)
+
+## [0.12.1]
+### Changed
+- Improved documentation on `Map::layers` and `Map::get_layer`. (#306)
+- Extend lifetime of `Layer`s returned from `GroupLayer::get_layer` to the map's. (#307)
+
 ## [0.12.0]
 ### Added
 - Add `text`, `width` and `height` members to `ObjectShape::Text`. (#278)